Refactor: rename module_ext.py to program_settings.py and update imports
