+++ esp/esp/program/controllers/confirmation.py
@@ -35,7 +35,7 @@ Learning Unlimited, Inc.
 
 from esp.program.models import Program, ClassSection, ClassSubject
 from esp.users.models import ESPUser, Record, RecordType
-from esp.program.modules.module_ext import DBReceipt
+from esp.program.modules.program_settings import DBReceipt
 from esp.program.modules.forms.admincore import get_template_source
 
 from django.template import Template, Context
