+++ esp/esp/program/modules/forms/teacherreg.py
@@ -41,7 +41,7 @@ from esp.utils.forms import StrippedCharField, FormWithRequiredCss, FormUnrestri
 from esp.utils.widgets import BlankSelectWidget, SplitDateWidget
 import re
 from esp.program.models import ClassCategories, ClassSubject, ClassSection, ClassSizeRange
-from esp.program.modules.module_ext import ClassRegModuleInfo
+from esp.program.modules.program_settings import ClassRegModuleInfo
 from esp.users.models import UserAvailability
 from esp.cal.models import Event
 from esp.tagdict.models import Tag
