+++ esp/esp/program/modules/models.py
@@ -111,4 +111,4 @@ def install(model=None):
 
     updateModules(table_data, deleteExtra=True, model=model)
 
-from esp.program.modules.module_ext import *
+from esp.program.modules.program_settings import *
