+++ esp/esp/program/modules/signals.py
@@ -1,5 +1,5 @@
 from esp.program.models import maybe_create_module_ext
-from esp.program.modules.module_ext import StudentClassRegModuleInfo, ClassRegModuleInfo
+from esp.program.modules.program_settings import StudentClassRegModuleInfo, ClassRegModuleInfo
 
 # TODO(benkraft): There are actually a lot more modules that depend on these
 # module extensions.  In practice it's probably fine because very few programs
